refactor: directory 변경 적용
