+++ src/test/java/com/pyro/yolog/domain/trip/api/TripControllerTest.java
@@ -2,7 +2,7 @@ package com.pyro.yolog.domain.trip.api;
 
 import com.pyro.yolog.domain.global.BaseControllerTest;
 import com.pyro.yolog.domain.trip.TripFixture;
-import com.pyro.yolog.domain.trip.dto.TripRequest;
+import com.pyro.yolog.domain.trip.dto.request.TripRequest;
 import com.pyro.yolog.domain.trip.entity.Trip;
 import com.pyro.yolog.domain.trip.repository.TripRepository;
 import com.pyro.yolog.domain.trip.service.TripService;
